requirements.txt
anthropic>=0.92.0
feedparser>=6.0.11
httpx>=0.27.0
playwright>=1.44.0
python-dotenv>=1.0.1
pyperclip>=1.9.0
matplotlib>=3.8.0
yfinance>=0.2.40
